+++ src/models/geometric_models.jl
@@ -187,7 +187,7 @@ of the Fourier expansion. The `N` in the type defines the order of
 the Fourier expansion.
 
 """
-function MRing(c::Union{AbstractVector{<:Complex}, NTuple{N,<:Complex}})
+function MRing(c::Union{AbstractVector{<:Complex}, NTuple{N,<:Complex}}) where {N}
     α = real.(c)
     β = imag.(c)
     return MRing(α, β)
